US01 detect center remap during sync
