+++ assets/l10n/app_en.arb
@@ -189,7 +189,7 @@
   "appliction_upload": "Application Upload",
   "check_updates": "Check Updates",
   "center_remap_sync": "Center Remap Sync.",
-  "centre_remap_notification": "Your machine has been re-mapped to another center. Kindly complete the pending activities.",
+  "center_remap_notification": "Your machine has been re-mapped to another center. Kindly complete the pending activities.",
   "sync_activities": "Sync. Activities",
   "pending_approval": "Pending Approval",
   "dashboard": "Dashboard",
