+++ assets/l10n/app_fr.arb
@@ -188,7 +188,7 @@
   "appliction_upload": "Téléchargement d'application",
   "check_updates": "Vérifier les mises à jour",
   "center_remap_sync": "Synchronisation du remappage central.",
-  "centre_remap_notification": "Votre machine a été réaffectée à un autre centre. Veuillez compléter les activités en attente.",
+  "center_remap_notification": "Votre machine a été réaffectée à un autre centre. Veuillez compléter les activités en attente.",
   "sync_activities": "Synchronisation. Activités",
   "pending_approval": "En attente de validation",
   "dashboard": "Tableau de bord",
