+++ assets/l10n/app_hi.arb
@@ -188,7 +188,7 @@
   "appliction_upload": "आवेदन अपलोड करें",
   "check_updates": "अपडेट जांचें",
   "center_remap_sync": "केंद्र रीमैप सिंक.",
-  "centre_remap_notification": "आपकी मशीन को किसी अन्य केंद्र पर पुनः मैप किया गया है। कृपया लंबित गतिविधियाँ पूरी करें।",
+  "center_remap_notification": "आपकी मशीन को किसी अन्य केंद्र पर पुनः मैप किया गया है। कृपया लंबित गतिविधियाँ पूरी करें।",
   "sync_activities": "साथ-साथ करना। गतिविधियाँ",
   "pending_approval": "लंबित अनुमोदन",
   "dashboard": "डैशबोर्ड",
