+++ assets/l10n/app_kn.arb
@@ -188,7 +188,7 @@
   "appliction_upload": "ಅಪ್ಲಿಕೇಶನ್ ಅಪ್ಲೋಡ್",
   "check_updates": "ನವೀಕರಣಗಳನ್ನು ಪರಿಶೀಲಿಸಿ",
   "center_remap_sync": "ಕೇಂದ್ರ ರೀಮ್ಯಾಪ್ ಸಿಂಕ್.",
-  "centre_remap_notification": "ನಿಮ್ಮ ಯಂತ್ರವನ್ನು ಇನ್ನೊಂದು ಕೇಂದ್ರಕ್ಕೆ ಮರು-ಮ್ಯಾಪ್ ಮಾಡಲಾಗಿದೆ. ದಯವಿಟ್ಟು ಬಾಕಿ ಇರುವ ಚಟುವಟಿಕೆಗಳನ್ನು ಪೂರ್ಣಗೊಳಿಸಿ.",
+  "center_remap_notification": "ನಿಮ್ಮ ಯಂತ್ರವನ್ನು ಇನ್ನೊಂದು ಕೇಂದ್ರಕ್ಕೆ ಮರು-ಮ್ಯಾಪ್ ಮಾಡಲಾಗಿದೆ. ದಯವಿಟ್ಟು ಬಾಕಿ ಇರುವ ಚಟುವಟಿಕೆಗಳನ್ನು ಪೂರ್ಣಗೊಳಿಸಿ.",
   "sync_activities": "ಸಿಂಕ್ ಮಾಡಿ. ಚಟುವಟಿಕೆಗಳು",
   "pending_approval": "ಒಪ್ಪಿಗೆಗಾಗಿ ಕಾದಿರುವ",
   "dashboard": "ಡ್ಯಾಶ್‌ಬೋರ್ಡ್",
