+++ assets/l10n/app_ta.arb
@@ -180,7 +180,7 @@
   "appliction_upload": "விண்ணப்பப் பதிவேற்றம்",
   "check_updates": "புதுப்பிப்புகளைச் சரிபார்க்கவும்",
   "center_remap_sync": "மைய ரீமேப் ஒத்திசைவு.",
-  "centre_remap_notification": "உங்கள் இயந்திரம் மற்றொரு மையத்திற்கு மீண்டும் மேப் செய்யப்பட்டது. நிலுவையிலுள்ள செயல்பாடுகளை முடிக்கவும்.",
+  "center_remap_notification": "உங்கள் இயந்திரம் மற்றொரு மையத்திற்கு மீண்டும் மேப் செய்யப்பட்டது. நிலுவையிலுள்ள செயல்பாடுகளை முடிக்கவும்.",
   "sync_activities": "ஒத்திசை செயல்பாடுகள்",
   "pending_approval": "நிலுவையிலுள்ள ஒப்புதல்",
   "dashboard": "டாஷ்போர்டு",
