+++ lib/ui/onboard/portrait/registration_tasks.dart
@@ -42,15 +42,53 @@ class _RegistrationTasksState extends State<RegistrationTasks> {
     super.initState();
   }
 
+  Widget _buildRemapBanner(BuildContext context) {
+    return Padding(
+      padding: EdgeInsets.symmetric(horizontal: 16.w, vertical: 6.h),
+      child: Container(
+        padding: EdgeInsets.symmetric(horizontal: 12.w, vertical: 14.h),
+        decoration: BoxDecoration(
+          color: const Color(0xFFFFFBE6),
+          border: Border.all(color: const Color(0xFFE6A817), width: 1.5),
+          borderRadius: BorderRadius.circular(8),
+        ),
+        child: Row(
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: [
+            Icon(
+              Icons.warning_amber_rounded,
+              color: const Color(0xFFE6A817),
+              size: isMobileSize ? 22 : 30,
+            ),
+            SizedBox(width: 10.w),
+            Expanded(
+              child: Text(
+                AppLocalizations.of(context)!.center_remap_notification,
+                style: TextStyle(
+                  color: const Color(0xFFE6A817),
+                  fontWeight: FontWeight.bold,
+                  fontSize: isMobileSize ? 13 : 18,
+                ),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
     isPortrait = MediaQuery.of(context).orientation == Orientation.portrait;
+    final bool centreRemapped =
+        context.watch<SyncProvider>().isCentreRemapped;
     return SingleChildScrollView(
       child: Column(
         children: [
           SizedBox(
             height: 26.h,
           ),
+          if (centreRemapped) _buildRemapBanner(context),
           isMobileSize
               ? Padding(
             padding: EdgeInsets.symmetric(horizontal: 16.w),
